+++ jdaviz/configs/cubeviz/plugins/spectral_extraction/tests/test_spectral_extraction.py
@@ -356,3 +356,16 @@ def test_cone_aperture_with_frequency_units(cubeviz_helper, spectral_cube_wcs):
 
     with pytest.raises(ValueError, match="Spectral axis unit physical type is"):
         extract_plg.collapse_to_spectrum()
+
+
+def test_cube_extraction_with_nan(cubeviz_helper, image_cube_hdu_obj):
+    image_cube_hdu_obj[1].data[:, :2, :2] = np.nan
+    cubeviz_helper.load_data(image_cube_hdu_obj, data_label="with_nan")
+    extract_plg = cubeviz_helper.plugins['Spectral Extraction']
+    sp = extract_plg.collapse_to_spectrum()  # Default settings (sum)
+    assert_allclose(sp.flux.value, 96)  # (10 x 10) - 4
+
+    cubeviz_helper.load_regions(RectanglePixelRegion(PixCoord(1.5, 1.5), width=4, height=4))
+    extract_plg.aperture = 'Subset 1'
+    sp_subset = extract_plg.collapse_to_spectrum()  # Default settings but on Subset
+    assert_allclose(sp_subset.flux.value, 12)  # (4 x 4) - 4